server/routes/deactivateAlertCodeRoutes.ts
import { RequestHandler } from 'express'
import AlertsApiClient from '../data/alertsApiClient'

export default class DeactivateAlertCodeRoutes {
  constructor(private readonly alertsApiClient: AlertsApiClient) {}

  public startPage: RequestHandler = async (req, res): Promise<void> => {
    const alertTypes = (await this.alertsApiClient.retrieveAlertTypes(req.middleware.clientToken)).map(alertType => {
      return {
        value: alertType.code,
        text: alertType.code,
        hint: { text: alertType.description },
      }
    })
    return res.render('pages/deactivateAlertCode/index', { alertTypes })
  }
}
